requirements.txt
bcrypt==4.0.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
Flask==3.0.0
Flask-Bcrypt==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
Pillow==10.1.0
pycparser==2.21
Werkzeug==3.0.1
